# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Construction Coordination Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Projects",
        "operationId": "projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "type": "array",
                  "title": "Response Projects Api Projects Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/workspace": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Workspace",
        "operationId": "workspace_api_projects__project_id__workspace_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/events": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Ingest",
        "operationId": "ingest_api_projects__project_id__events_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectEvent-Input"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/recheck": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Recheck",
        "operationId": "recheck_api_projects__project_id__recheck_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/reset": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Reset Demo",
        "operationId": "reset_demo_api_demo_reset_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/runs": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Project Runs",
        "operationId": "project_runs_api_projects__project_id__runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentRun"
                  },
                  "title": "Response Project Runs Api Projects  Project Id  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActionProposal": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Generation",
            "default": 0
          },
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "snapshot_version": {
            "type": "integer",
            "title": "Snapshot Version"
          },
          "work_package_id": {
            "type": "string",
            "title": "Work Package Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "risk": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Risk"
          },
          "resolution": {
            "$ref": "#/components/schemas/ResolutionOption"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "simulated",
              "external"
            ],
            "title": "Execution Mode",
            "default": "simulated"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "operation_id",
          "project_id",
          "run_id",
          "generation",
          "snapshot_id",
          "snapshot_version",
          "work_package_id",
          "title",
          "risk",
          "resolution",
          "evidence_ids",
          "execution_mode",
          "created_at"
        ],
        "title": "ActionProposal"
      },
      "AgentRun": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "WAITING_APPROVAL",
              "COMPLETED",
              "FAILED",
              "CANCELLED",
              "EXPIRED"
            ],
            "title": "Status",
            "default": "QUEUED"
          },
          "runtime": {
            "type": "string",
            "title": "Runtime",
            "default": "dbos"
          },
          "runtime_execution_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Execution Id"
          },
          "generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Generation",
            "default": 0
          },
          "runtime_generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Runtime Generation",
            "default": 0
          },
          "category": {
            "type": "string",
            "enum": [
              "coordination",
              "document_parse",
              "bim_import",
              "optimization",
              "vision",
              "embedding_index"
            ],
            "title": "Category",
            "default": "coordination"
          },
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "event_id",
          "status",
          "runtime",
          "runtime_execution_id",
          "generation",
          "runtime_generation",
          "category",
          "analysis_id",
          "error",
          "created_at",
          "updated_at"
        ],
        "title": "AgentRun"
      },
      "Analysis": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "snapshot": {
            "$ref": "#/components/schemas/ProjectSnapshot"
          },
          "impact": {
            "$ref": "#/components/schemas/Impact"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/Constraint"
            },
            "type": "array",
            "title": "Constraints"
          },
          "readiness": {
            "items": {
              "$ref": "#/components/schemas/Readiness"
            },
            "type": "array",
            "title": "Readiness"
          },
          "reasoning_summary": {
            "type": "string",
            "title": "Reasoning Summary"
          },
          "reasoning_mode": {
            "type": "string",
            "title": "Reasoning Mode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "snapshot",
          "impact",
          "evidence",
          "findings",
          "constraints",
          "readiness",
          "reasoning_summary",
          "reasoning_mode",
          "created_at"
        ],
        "title": "Analysis"
      },
      "Approval": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "principal_id": {
            "type": "string",
            "title": "Principal Id"
          },
          "principal_role": {
            "type": "string",
            "title": "Principal Role"
          },
          "level": {
            "type": "string",
            "enum": [
              "standard",
              "strong"
            ],
            "title": "Level"
          },
          "confirmation": {
            "type": "string",
            "title": "Confirmation",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "proposal_id",
          "principal_id",
          "principal_role",
          "level",
          "confirmation",
          "created_at"
        ],
        "title": "Approval"
      },
...
      "Area": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "floor": {
            "type": "string",
            "title": "Floor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "floor"
        ],
        "title": "Area"
      },
      "AuditRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "detail": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Detail"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "action",
          "actor",
          "run_id",
          "snapshot_id",
          "operation_id",
          "detail",
          "created_at"
        ],
        "title": "AuditRecord"
      },
...
      "EventChange-Input": {
        "properties": {
          "revision": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "available_workers": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Workers"
          },
          "qualifications": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualifications"
          },
          "predecessor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predecessor Id"
          },
          "complete": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complete"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available"
          },
          "inspection_passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspection Passed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EventChange"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "timezone"
        ],
        "title": "Project"
      },
      "ProjectEvent-Input": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "design_revision",
              "workforce",
              "predecessor",
              "material",
              "equipment",
              "inspection",
              "external"
            ],
            "title": "Kind"
          },
          "work_package_id": {
            "type": "string",
            "title": "Work Package Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "note": {
            "type": "string",
            "maxLength": 4000,
            "title": "Note",
            "default": ""
          },
          "source": {
            "type": "string",
            "maxLength": 200,
            "title": "Source",
            "default": "local-demo"
          },
          "change": {
            "$ref": "#/components/schemas/EventChange-Input"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "kind",
          "work_package_id",
          "title"
        ],
        "title": "ProjectEvent"
      },
      "ProjectEvent-Output": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "design_revision",
              "workforce",
              "predecessor",
              "material",
              "equipment",
              "inspection",
              "external"
            ],
            "title": "Kind"
          },
          "work_package_id": {
            "type": "string",
            "title": "Work Package Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "note": {
            "type": "string",
            "maxLength": 4000,
            "title": "Note",
            "default": ""
          },
          "source": {
            "type": "string",
            "maxLength": 200,
            "title": "Source",
            "default": "local-demo"
          },
          "change": {
            "$ref": "#/components/schemas/EventChange-Output"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "kind",
          "work_package_id",
          "title",
          "note",
          "source",
          "change",
          "observed_at"
        ],
        "title": "ProjectEvent"
      },
...
      "ProjectState": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "default": 1
          },
          "areas": {
            "items": {
              "$ref": "#/components/schemas/Area"
            },
            "type": "array",
            "title": "Areas"
          },
          "work_packages": {
            "items": {
              "$ref": "#/components/schemas/WorkPackage"
            },
            "type": "array",
            "title": "Work Packages"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRevision"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project",
          "version",
          "areas",
          "work_packages",
          "sources"
        ],
        "title": "ProjectState"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkPackage": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "area_id": {
            "type": "string",
            "title": "Area Id"
          },
          "discipline": {
            "type": "string",
            "title": "Discipline"
          },
          "element_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Element Ids",
            "default": []
          },
          "predecessors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Predecessors",
            "default": []
          },
          "complete": {
            "type": "boolean",
            "title": "Complete",
            "default": false
          },
          "design_revision": {
            "type": "string",
            "title": "Design Revision",
            "default": "V16"
          },
          "accepted_revision": {
            "type": "string",
            "title": "Accepted Revision",
            "default": "V16"
          },
          "required_workers": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Required Workers",
            "default": 2
          },
          "available_workers": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Available Workers",
            "default": 2
          },
          "required_qualifications": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Qualifications",
            "default": []
          },
          "qualifications": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Qualifications",
            "default": []
          },
          "materials": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Materials"
          },
          "equipment": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Equipment"
          },
          "inspection_passed": {
            "type": "boolean",
            "title": "Inspection Passed",
            "default": true
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": "Coordination team"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "area_id",
          "discipline",
          "element_ids",
          "predecessors",
          "complete",
          "design_revision",
          "accepted_revision",
          "required_workers",
          "available_workers",
          "required_qualifications",
          "qualifications",
          "materials",
          "equipment",
          "inspection_passed",
          "owner"
        ],
        "title": "WorkPackage"
      },
      "WorkspaceResponse": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/ProjectState"
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Analysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "analysis_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/ActionProposal"
            },
            "type": "array",
            "title": "Proposals"
          },
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/Approval"
            },
            "type": "array",
            "title": "Approvals"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ProjectEvent-Output"
            },
            "type": "array",
            "title": "Events"
          },
          "audit": {
            "items": {
              "$ref": "#/components/schemas/AuditRecord"
            },
            "type": "array",
            "title": "Audit"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "state",
          "analysis",
          "run",
          "analysis_run",
          "proposals",
          "approvals",
          "events",
          "audit",
          "stale"
        ],
        "title": "WorkspaceResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}